Pre-final version of the tests(only apigateway is left)
